Category, Tagテーブルにcount追加して記事数を保持するロジックを実装
メール機能追加

+++ src/app/admin/posts/actions.ts
@@ -23,7 +23,17 @@ export const createPost = adminActionClient.inputSchema(ZPost).action(async ({ p
       authorId: parsedInput.authorId,
     };
 
-    return await prisma.post.create({ data: saveData });
+    return await prisma.$transaction([
+      prisma.post.create({ data: saveData }),
+      prisma.category.update({
+        where: { id: parsedInput.categoryId },
+        data: { count: { increment: 1 } },
+      }),
+      prisma.tag.updateMany({
+        where: { id: { in: parsedInput.tagIds } },
+        data: { count: { increment: 1 } },
+      }),
+    ]);
   } catch (e) {
     // todo エラー処理 存在しないタグIDを指定した場合は例外がスローされる
     if (e instanceof Prisma.PrismaClientKnownRequestError) {
@@ -41,25 +51,68 @@ export const updatePost = adminActionClient.inputSchema(ZPost).action(async ({ p
       throw new ResourceNotFoundError('Post', null);
     }
 
+    const newTagIds = parsedInput.tagIds.sort();
+    const oldTagIds = post.tags.map((tag) => tag.id).sort();
+    const addedTagIds = newTagIds.filter((i) => oldTagIds.indexOf(i) == -1);
+    const deleteTagIds = oldTagIds.filter((i) => newTagIds.indexOf(i) == -1);
+
     const saveData = {
       title: parsedInput.title,
       content: parsedInput.content,
       published: parsedInput.published,
       categoryId: parsedInput.categoryId,
       authorId: parsedInput.authorId,
-      tags: {
-        // 存在しないタグIDを指定した場合は例外がスローされる
-        set: parsedInput.tagIds.map((id) => {
+    } as Prisma.PostUpdateInput;
+
+    if (newTagIds.toString() !== oldTagIds.toString()) {
+      saveData.tags = {
+        set: newTagIds.map((id) => {
           return { id };
         }),
-      },
-    };
+      };
+    }
 
-    return await prisma.post.update({
-      where: { id: id },
-      data: saveData,
-    });
+    const transactionArray: any = [
+      prisma.post.update({
+        where: { id: id },
+        data: saveData,
+      }),
+    ];
+    const isCategoryChanged = post.categoryId !== parsedInput.categoryId;
+    if (isCategoryChanged) {
+      transactionArray.push(
+        prisma.category.update({
+          where: { id: parsedInput.categoryId },
+          data: { count: { increment: 1 } },
+        }),
+        prisma.category.update({
+          where: { id: post.categoryId },
+          data: { count: { decrement: 1 } },
+        }),
+      );
+    }
+
+    if (addedTagIds) {
+      transactionArray.push(
+        prisma.tag.updateMany({
+          where: { id: { in: addedTagIds } },
+          data: { count: { increment: 1 } },
+        }),
+      );
+    }
+
+    if (deleteTagIds) {
+      transactionArray.push(
+        prisma.tag.updateMany({
+          where: { id: { in: deleteTagIds } },
+          data: { count: { decrement: 1 } },
+        }),
+      );
+    }
+
+    return await prisma.$transaction(transactionArray);
   } catch (e) {
+    // 存在しないタグIDを指定した場合は例外がスローされる
     if (e instanceof Prisma.PrismaClientKnownRequestError) {
       throw new DatabaseError(e.message);
     }
@@ -71,7 +124,25 @@ const ZDeletePost = z.object({ id: ZCuid });
 
 export const deletePost = adminActionClient.inputSchema(ZDeletePost).action(async ({ parsedInput }) => {
   try {
-    await prisma.post.delete({ where: { id: parsedInput.id } });
+    const post = await prisma.post.findUnique({
+      include: { tags: true },
+      where: { id: parsedInput.id },
+    });
+    if (!post) {
+      throw new ResourceNotFoundError('Post', null);
+    }
+
+    await prisma.$transaction([
+      prisma.post.delete({ where: { id: parsedInput.id } }),
+      prisma.category.update({
+        where: { id: post.categoryId },
+        data: { count: { decrement: 1 } },
+      }),
+      prisma.tag.updateMany({
+        where: { id: { in: post.tags.map((tag) => tag.id) } },
+        data: { count: { decrement: 1 } },
+      }),
+    ]);
     return true;
   } catch (e) {
     if (e instanceof Prisma.PrismaClientKnownRequestError) {
